Fixed exercise 4.4
