+++ fpinkotlin-parent/fpinkotlin-recursion-exercises/src/main/kotlin/com/fpinkotlin/recursion/exercise04/MakeString.kt
@@ -1,0 +1,17 @@
+package com.fpinkotlin.recursion.exercise04
+
+fun <T> makeString(list: List<T>, delim: String): String {
+    TODO("makeString")
+}
+
+fun <T> List<T>.head(): T =
+        if (this.isEmpty())
+            throw IllegalArgumentException("head called on empty list")
+        else
+            this[0]
+
+fun <T> List<T>.tail(): List<T> =
+        if (this.isEmpty())
+            throw IllegalArgumentException("tail called on empty list")
+        else
+            this.subList(1, this.size)
