+++ fpinkotlin-parent/fpinkotlin-recursion-exercises/src/main/kotlin/com/fpinkotlin/recursion/exercise04/String.kt
@@ -1,15 +1,3 @@
 package com.fpinkotlin.recursion.exercise04
 
 fun string(list: List<Char>): String = TODO("string")
-
-fun <T> List<T>.head(): T =
-        if (this.isEmpty())
-            throw IllegalArgumentException("head called on empty list")
-        else
-            this[0]
-
-fun <T> List<T>.tail(): List<T> =
-        if (this.isEmpty())
-            throw IllegalArgumentException("tail called on empty list")
-        else
-            this.subList(1, this.size)
